Conflicts:
	src/main/java/net/herdao/hdp/manpower/mpclient/service/impl/RecruitmentServiceImpl.java

+++ src/main/java/net/herdao/hdp/manpower/mpclient/dto/staffChanges/SaveStaffCallOutDTO.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 /**
  * @Author Liu Chang
@@ -56,13 +57,13 @@ public class SaveStaffCallOutDTO {
     private Boolean isPrepareChange;
 
     @ApiModelProperty(value = "合同年限", name = "contractPeriod")
-    private Double contractPeriod;
+    private BigDecimal contractPeriod;
 
     @ApiModelProperty(value = "本公司工龄", name = "companySeniority")
-    private Double companySeniority;
+    private BigDecimal companySeniority;
 
     @ApiModelProperty(value = "合生珠江系工龄", name = "threeGroupSeniority")
-    private Double threeGroupSeniority;
+    private BigDecimal threeGroupSeniority;
 
     @ApiModelProperty(value = "经办人意见", name = "remark")
     private String remark;
